+++ src/ezpr/index.ts
@@ -2,6 +2,6 @@ export * from "./";
 export * from "./args";
 export * from "./blocks";
 export * from "./cmd";
-export * from "./form_submission";
+export * from "./formSubmission";
 export * from "./listeners";
-export * from "./slash_command";
+export * from "./slashCommand";
