+++ src/help/listeners/openHelpModal.ts
