+++ src/help/slashCommand.ts
@@ -1,6 +1,6 @@
 import { SlashCommand } from "@slack/bolt";
 
-import { HelpArguments } from "./";
+import { HelpArguments } from ".";
 
 export function ParseSlashHelpCommand(payload: SlashCommand): HelpArguments {
   const input = `${payload.command} ${payload.text}`;
